rename files templatepreprocessor -> handlebarspreprocessor
