+++ src/handlebars/handlebarspreprocessor.js
