+++ tests/fixtures/handlebars/processedcomponent.js
