+++ tests/fixtures/handlebars/rawcomponent.js
