+++ tests/fixtures/handlebars/rawtemplate.hbs
